### 2025-02-17 #686 @smnnlt
c3dr: Read and Write C3D Motion Capture Data in R
Submitting Author Name: Simon Nolte
Submitting Author Github Handle: @smnnlt
Repository:  https://github.com/smnnlt/c3dr
Version submitted: 0.1.0
Submission type: Standard

Version accepted: TBD

---

```
Package: c3dr
Title: Read and Write C3D Motion Capture Files
Version: 0.1.0.9000
    person("Simon", "Nolte", , "s.nolte@dshs-koeln.de", role = c("aut", "cre"),
Description: An wrapper for the ‘EZC3D’ library to work with C3D motion capture data.
    Rcpp
    Rcpp
    quarto
    quarto
```

The c3dr package allows to read, process, and write biomechanical data in the C3D format (e.g., motion capture data).

The package is written for anyone working in biomechanics. C3D is the standard file format for motion capture analysis.

There are no other R packages, only Python, Matlab, or C++ libraries.

n/a

I couldn't run pkgcheck locally because of ropensci-review-tools/pkgstats/issues/65
