### 2025-02-17 #686 @mpadge
Sorry @smnnlt, You'll see from the above PRs that I'm updating our system to include the quarto binary so your package will pass. It's currently offline, but I'll ping here again as soon as it's back up. In the meantime, you could install our [pkgcheck-action](https://github.com/ropensci-review-tools/pkgcheck-action) in your repo, and confirm checks with each local commit that way. I've already bundled the quarto binary installation there, so that should be good to go.
