### 2025-02-18 #686 @maurolepore
@smnnlt thanks so much for sharing your work with rOpenSci.

Here's a quick preliminary check to support the discussion with the editorial board. We'll be back here ASAP.

  > TBD
  > Copy-paste just worked :-)
  > NA
  > The initial commit was mid-last years but I see no issue or PR (neither open or closed). The commit history is very clear. This suggests a mature workflow, not based on GitHub PRs.

---

* Dependencies: Only Rcpp
* Size: Appropriate
* Style: Seems to follow best practices. In the roxygen documentation I see a syntax that could be a lil less verbose:
```
# I see:
#' @return A data frame of class \code{c3d_data} in the format specified by the
#'   \code{format} argument.

# Could be
#' @return A data frame of class `c3d_data` in the format specified by the
#'   `format` argument.
```

---
