### 2025-02-18 #686 @smnnlt
@maurolepore Thank you for this first review. Based on your helpful comment, I have worked in the [dev](smnnlt/c3dr/tree/dev) branch to convert the documentation to Markdown syntax. I have also shortened one line of code as indicated by the lintr check.
