### 2025-02-18 #686 @maurolepore
Thanks @smnnlt, it's great to see you're kind, open to feedback and responsive. That's an attitude that makes us all feel great about contributing out time to advance software in research.

While you're there you might also want to consider this suggestion that just came from the editorial board:

> They include the original MIT license file in src/. It'd be good to ask them to attribute in DESCRIPTION with cph (as it's MIT-licensed). Here's a link for them: https://r-pkgs.org/license.html#sec-how-to-include
