### 2025-02-18 #686 @smnnlt
@maurolepore Thank you for the kind words!

I have updated the description file as suggested https://github.com/smnnlt/c3dr/commit/f2caea4d5ebe669645d0f364d27aa873b4feb887.
