### 2025-02-25 #686 @maelle
---

Thank you for your submission! I only have these few comments that should be tackled (or ignored, as some are suggestions!) before I start looking for reviewers.

##### Metadata

If the German Sport University Cologne has a ROR ID, you could add it to the DESCRIPTION file.
If https://ror.org/0189raq88 is the right one,

```r
pak::pak("r-lib/desc") # only in the dev version at the moment
desc::desc_add_author("German Sport University Cologne", role = "fnd", ror = "0189raq88")

```

##### README

- In the [authors guide](https://devguide.ropensci.org/softwarereview_author.html), we recommend:

> If you use repostatus.org badges (which we recommend), submit when you’re ready to get an Active instead of WIP badge.

Could you please update the badge in the README?

- You could recommend using pak rather than remotes, see for instance [usethis README](https://github.com/r-lib/usethis/?tab=readme-ov-file#installation)

```r
# install.packages("pak")
pak::pak("smnnlt/c3dr")
```

- Seeing the acknowledgements section reminded of the [allcontributors R package](https://ropensci.org/blog/2024/11/26/allcontributors/) that you might find interesting.

##### Contributing guide

I like how it's structured around different use cases, it looks very actionable. :smile_cat:

https://github.com/smnnlt/c3dr/blob/main/.github/CONTRIBUTING.md#share-the-work mentions Twitter, which is no longer Twitter, and not necessarily the friendliest place ever.
I'd suggest updating it to "social media".

##### roxygen2

You have turned on Markdown support in `DESCRIPTION`(`Roxygen: list(markdown = TRUE)`) but you use `\code{}` and `\link{}` instead of their [Markdown equivalents](https://roxygen2.r-lib.org/articles/rd-formatting.html).
You don't have to switch, but switching might make maintenance of roxygen2 content easier.

##### Code

- No obligation at all, but I wanted to mention that messages like `stop("'x' needs to be a list of class 'c3d'.")` might be easier to format with [cli](https://cli.r-lib.org/). `cli::cli_abort("{.arg x} must be a list of class {.cls c3d}, not {.obj_type_friendly {x}}.")`

- If you keep using `stop()` (and `warning()`, `message()`),

```r
    stop(
      paste0(
        "'data' needs to be a data frame of class",
        ") 'c3d_data' as generated by 'cd3_data()'."
      )
    )
```

can be

```r
    stop(
        "'data' needs to be a data frame of class",
        ") 'c3d_data' as generated by 'cd3_data()'."
    )
```

(I had forgotten about this, [lintr](https://masalmon.eu/2024/08/28/lintr-3-steps/) reminded me!)
By the way is there a stray parenthesis here?

- Also caught by lintr, is this variable used: https://github.com/smnnlt/c3dr/blob/b435c3e412e1f595cfae2ad5b2fdc16d3d2657f2/R/convert.R#L192

- Is the `c()` needed in https://github.com/smnnlt/c3dr/blob/b435c3e412e1f595cfae2ad5b2fdc16d3d2657f2/R/convert.R#L266?

- Yet again courtesy of lintr,

```r
  # convert to wide data format if necessary
  if (!inherits(newdata, "c3d_data_wide")) {
    d <- c3d_convert(newdata, "wide")
    d <- newdata
  }
```

might be more readable as

```r
  # convert to wide data format if necessary
  if (inherits(newdata, "c3d_data_wide")) {
    d <- newdata
    d <- c3d_convert(newdata, "wide")
  }
```

- R CMD check finds a NOTE that I do not know about (given I never use C++ :sweat_smile:) Is it a matter of metadata somewhere?

```
❯ checking if this is a source package ... NOTE
  Found the following apparent object files/libraries:
    src/ezc3d/Analogs.o src/ezc3d/AnalogsInfo.o
    src/ezc3d/AnalogsSubframe.o src/ezc3d/Channel.o src/ezc3d/Data.o
    src/ezc3d/DataStartInfo.o src/ezc3d/ezc3d.o src/ezc3d/Frame.o
    src/ezc3d/Group.o src/ezc3d/Header.o src/ezc3d/math/Matrix.o
    src/ezc3d/math/Matrix33.o src/ezc3d/math/Matrix44.o
    src/ezc3d/math/Matrix66.o src/ezc3d/math/Vector3d.o
    src/ezc3d/math/Vector6d.o src/ezc3d/modules/ForcePlatforms.o
    src/ezc3d/Parameter.o src/ezc3d/Parameters.o src/ezc3d/Point.o
    src/ezc3d/Points.o src/ezc3d/PointsInfo.o src/ezc3d/Rotation.o
    src/ezc3d/Rotations.o src/ezc3d/RotationsInfo.o
    src/ezc3d/RotationsSubframe.o
  Object files/libraries should not be included in a source package.
```

##### Docs

- When using the example dataset, maybe not in the README but in the vignette at least, could you add a sentence or two about what's been measured in it?

- What's the next step for users, will they plot the data? I wonder whether adding a small example of this in the vignette would help.

- Should the vignette define "point data" and "analog data"?

- A suggestion for the vignette. For future Git diffs, you might want to switch to adding a line break after each sentence. That's a nitpick of mine so here's a package to help: <https://docs.ropensci.org/aeolus/>

##### Tests

- In test files you have "top-level code" such as

```r
# get data
d <- c3d_read(c3d_example())

d_wide <- c3d_data(d)
d_long <- c3d_data(d, format = "long")
d_longest <- c3d_data(d, format = "longest")
a <- c3d_analog(d)
```

I would recommend to instead create the data again _within each test_.
This way, if a a test fails, it's easier to know where the data comes from (as opposed to scrolling up).
Everything needed within a `test_that()` call should be defined within it (using, if needed, a call to a function defined in a [helper file](https://r-pkgs.org/testing-design.html#testthat-helper-files))
See <https://r-pkgs.org/testing-design.html#plan-for-test-failure> for more context on this.

- I'd recommend using `withr::local_tempfile()` within the test rather than these top-level lines:

```r
# create temporary file
tmp <- tempfile()
on.exit(unlink(tmp))
```

- The call to `expect_error()` should include the `regexp` or `class` argument to be more specific. I've never used classed errors myself but here's a neat [blog post by Mike Mahoney](https://www.mm218.dev/posts/2023-11-07-classed-errors/).

- Expectations like `expect_identical(class(d_wide), c("c3d_data_wide", "c3d_data", "data.frame"))` should be `expect_s3_class(d_wide, c("c3d_data_wide", "c3d_data", "data.frame"))`.

- Instead of `expect_identical(length())` you can use `expect_length()`.

---

Happy to answer any question/clarify/discuss. 😸
