### 2025-02-25 #686 @maelle
Oops sorry I see a bit too late @maurolepore had already commented on Markdown support in roxygen2 and that you're working on it!
