### 2025-02-27 #686 @smnnlt
@maelle Thank you very much for the your review! That was really helpful and I certainly learned a few new things. I made some modifcation and merged them into the main branch. Here is my response to the comments:

> ##### Metadata
>
> If the German Sport University Cologne has a ROR ID, you could add it to the DESCRIPTION file. If https://ror.org/0189raq88 is the right one,
>
> pak::pak("r-lib/desc") # only in the dev version at the moment
> desc::desc_add_author("German Sport University Cologne", role = "fnd", ror = "0189raq88")

Thanks! I have added my university with its ROR as a funder to the description https://github.com/smnnlt/c3dr/commit/4e7db681e958ea169921afc6a3e2ed342b75fd73
>
> ##### README
>
>  In the [authors guide](https://devguide.ropensci.org/softwarereview_author.html), we recommend:
>
>
> > If you use repostatus.org badges (which we recommend), submit when you’re ready to get an Active instead of WIP badge.
>
> Could you please update the badge in the README?

Done https://github.com/smnnlt/c3dr/commit/5da4bc8ce95e1eac93b9f310fccd5798d2c5e43d
>
>  You could recommend using pak rather than remotes, see for instance [usethis README](https://github.com/r-lib/usethis/?tab=readme-ov-file#installation)
>
>
> # install.packages("pak")
> pak::pak("smnnlt/c3dr")

I referenced pak before, but decided to switch to remotes instead. My main reason is that pak does not show any progress while compiling the C++ code (which remotes does), so users may get the impression that the installation is not working and terminate it early. Both pak and remotes had the same installation speed for me. So I do not see any benefits of recommending pak instead of remotes at the moment, but I am happy to be convinced otherwise. That said, I discovered that the vignette still recommended pak, so I changed that to remotes to be consistent https://github.com/smnnlt/c3dr/commit/0b5b3e6da3d42e7aa80be74fa7df2d546caaac63
>
> Seeing the acknowledgements section reminded of the [allcontributors R package](https://ropensci.org/blog/2024/11/26/allcontributors/) that you might find interesting.
>

I have discovered the allcontributors package some weeks ago and I think it is extremely helpful for some projects. As this project here currently only has a few (non-GitHub!) contributions, I do not see a value in using allcontributors. But I consider to change that later.
>
> ##### Contributing guide
>
> I like how it's structured around different use cases, it looks very actionable. 😸
>
> https://github.com/smnnlt/c3dr/blob/main/.github/CONTRIBUTING.md#share-the-work mentions Twitter, which is no longer Twitter, and not necessarily the friendliest place ever. I'd suggest updating it to "social media".

Good catch! https://github.com/smnnlt/c3dr/commit/c2ff0a46d65d4006a9e531a98c20d5b89a764329
> ##### roxygen2
>
> You have turned on Markdown support in `DESCRIPTION`(`Roxygen: list(markdown = TRUE)`) but you use `\code{}` and `\link{}` instead of their [Markdown equivalents](https://roxygen2.r-lib.org/articles/rd-formatting.html). You don't have to switch, but switching might make maintenance of roxygen2 content easier.

As you have noticed, I already switched to Markdown descriptions based on the suggestion by @maurolepore https://github.com/smnnlt/c3dr/commit/dd6ce9d8b240a7a537da133f7ba6daed43a781b1
> ##### Code
>
> No obligation at all, but I wanted to mention that messages like `stop("'x' needs to be a list of class 'c3d'.")` might be easier to format with [cli](https://cli.r-lib.org/). `cli::cli_abort("{.arg x} must be a list of class {.cls c3d}, not {.obj_type_friendly {x}}.")`

I agree that cli creates nicer messages. But I want to keep dependencies at the minimum, so I decided to stick to base R.
>
> If you keep using `stop()` (and `warning()`, `message()`),
>
>
>     stop(
>       paste0(
>         "'data' needs to be a data frame of class",
>         ") 'c3d_data' as generated by 'cd3_data()'."
>       )
>     )
>
> can be
>
>     stop(
>         "'data' needs to be a data frame of class",
>         ") 'c3d_data' as generated by 'cd3_data()'."
>     )
>
> (I had forgotten about this, [lintr](https://masalmon.eu/2024/08/28/lintr-3-steps/) reminded me!) By the way is there a stray parenthesis here?

Oh, I did not know that. I changed this in two instances (and there was indeed a stray parenthesis) https://github.com/smnnlt/c3dr/commit/b75c9a6d4da06f5f1a8314c959c580a92eeb8504
>
> Also caught by lintr, is this variable used: https://github.com/smnnlt/c3dr/blob/b435c3e412e1f595cfae2ad5b2fdc16d3d2657f2/R/convert.R#L192

Actually, no. https://github.com/smnnlt/c3dr/commit/4586791bf4e6d365d5d457f51394c5b898a9e4d0
>
> Is the `c()` needed in https://github.com/smnnlt/c3dr/blob/b435c3e412e1f595cfae2ad5b2fdc16d3d2657f2/R/convert.R#L266?

I realized it is not needed in some cases https://github.com/smnnlt/c3dr/commit/7f2e43ae55ce45e0d6fd27532159ed60856f5f40
>     * Yet again courtesy of lintr,
>
>
>   # convert to wide data format if necessary
>   if (!inherits(newdata, "c3d_data_wide")) {
>     d <- c3d_convert(newdata, "wide")
>   } else {
>     d <- newdata
>   }
>
> might be more readable as
>
>   # convert to wide data format if necessary
>   if (inherits(newdata, "c3d_data_wide")) {
>     d <- newdata
>   } else {
>     d <- c3d_convert(newdata, "wide")
>   }

Done https://github.com/smnnlt/c3dr/commit/8620e5ca82266ea01cfc0865012d3f4261a878b2
>
> R CMD check finds a NOTE that I do not know about (given I never use C++ 😅) Is it a matter of metadata somewhere?
>
>
> ```
> ❯ checking if this is a source package ... NOTE
>   Found the following apparent object files/libraries:
>     src/ezc3d/Analogs.o src/ezc3d/AnalogsInfo.o
>     src/ezc3d/AnalogsSubframe.o src/ezc3d/Channel.o src/ezc3d/Data.o
>     src/ezc3d/DataStartInfo.o src/ezc3d/ezc3d.o src/ezc3d/Frame.o
>     src/ezc3d/Group.o src/ezc3d/Header.o src/ezc3d/math/Matrix.o
>     src/ezc3d/math/Matrix33.o src/ezc3d/math/Matrix44.o
>     src/ezc3d/math/Matrix66.o src/ezc3d/math/Vector3d.o
>     src/ezc3d/math/Vector6d.o src/ezc3d/modules/ForcePlatforms.o
>     src/ezc3d/Parameter.o src/ezc3d/Parameters.o src/ezc3d/Point.o
>     src/ezc3d/Points.o src/ezc3d/PointsInfo.o src/ezc3d/Rotation.o
>     src/ezc3d/Rotations.o src/ezc3d/RotationsInfo.o
>     src/ezc3d/RotationsSubframe.o
>   Object files/libraries should not be included in a source package.
> ```

From how I understand it these are artifacts from the C++ code when building the repo locally. I get the same message on my local device, but the CI R-CMD does not show any notes.
> ##### Docs
>
> When using the example dataset, maybe not in the README but in the vignette at least, could you add a sentence or two about what's been measured in it?

Good idea! I added more context to the vignette https://github.com/smnnlt/c3dr/commit/fd6f065830a127d8dbe05469e15330884ab8f9ce
I also uploaded a video of the recording https://github.com/smnnlt/c3dr/blob/main/video/index.md that is referenced in the vignette. I hope this helps to better understand what the example dataset depicts.
>
> What's the next step for users, will they plot the data? I wonder whether adding a small example of this in the vignette would help.

The package is made for data import and processing, but not for further data analysis and visualization. However, I get that it is helpful to give such examples in the vignette. I added two examples of data visualization https://github.com/smnnlt/c3dr/commit/c6473b8fce7ce25fa96f2b1367cdc622e7c119ea
>
> Should the vignette define "point data" and "analog data"?

I tried to explain the terms better in the vignette https://github.com/smnnlt/c3dr/commit/56941138db1a6e8b2887750491b86fe29f5b4b5d
>
> A suggestion for the vignette. For future Git diffs, you might want to switch to adding a line break after each sentence. That's a nitpick of mine so here's a package to help: https://docs.ropensci.org/aeolus/

That is a really good idea! And the aeolus package was helpful for this, thank you for the suggestion https://github.com/smnnlt/c3dr/commit/df290ca0cf08ebb13a413cffca248375eb1d8348
>
>
> ##### Tests
>
>  In test files you have "top-level code" such as
>
>
> # get data
> d <- c3d_read(c3d_example())
>
> d_wide <- c3d_data(d)
> d_long <- c3d_data(d, format = "long")
> d_longest <- c3d_data(d, format = "longest")
> a <- c3d_analog(d)
>
> I would recommend to instead create the data again _within each test_. This way, if a a test fails, it's easier to know where the data comes from (as opposed to scrolling up). Everything needed within a `test_that()` call should be defined within it (using, if needed, a call to a function defined in a [helper file](https://r-pkgs.org/testing-design.html#testthat-helper-files)) See https://r-pkgs.org/testing-design.html#plan-for-test-failure for more context on this.

I did not consider this when writing the test code, because I felt it was against the principle of being not repetitive. But after I changed the test code, during the further test modifications you suggested, I realized how helpful it is to have all relevant code within the test function. So this was a really good suggestion! https://github.com/smnnlt/c3dr/commit/72eac7d731685c7ad3d1bbb00401e2f5671a2bbd
>
> I'd recommend using `withr::local_tempfile()` within the test rather than these top-level lines:
>
>
> # create temporary file
> tmp <- tempfile()
> on.exit(unlink(tmp))

I do not see any benefit compared to the base R version, despite the code being a bit cleaner, so I do not want to add a dependeny (even when only in Suggests) for this. (Though I just realized that testthat has withr as a dependency, so my argument may be stupid)
>
> The call to `expect_error()` should include the `regexp` or `class` argument to be more specific. I've never used classed errors myself but here's a neat [blog post by Mike Mahoney](https://www.mm218.dev/posts/2023-11-07-classed-errors/).

I did not know about classed errors, and I do not feel comfortable with using them in my package yet. Instead I used the `regexp` argument for all tests of messages, warnings and errors instead https://github.com/smnnlt/c3dr/commit/2390f2f04e0ab893943c6f69db4a562b6d601a99

And by the way, that was a good suggestion, as I discovered some messages that needed to be improved https://github.com/smnnlt/c3dr/commit/16359398f853d6cc04a04c84e5925ed8378953d8
>
> Expectations like `expect_identical(class(d_wide), c("c3d_data_wide", "c3d_data", "data.frame"))` should be `expect_s3_class(d_wide, c("c3d_data_wide", "c3d_data", "data.frame"))`.
>
>  Instead of `expect_identical(length())` you can use `expect_length()`.

Done for all cases! https://github.com/smnnlt/c3dr/commit/3a1fba5a8de173f5bfa50802ef815602b92596a9
> Happy to answer any question/clarify/discuss. 😸

Thank you very much for your review! Let me know if there is anything else I should adjust.
