### 2025-03-11 #686 @japilo
---

#### Tests and Checks

I did not mark that installation succeeds as documented because, while the package did install successfully, there were several warnings generated during the Rcpp installation. The warnings looked like this:

>   168 |   _LIBCPP_DEPRECATED_IN_CXX17 _LIBCPP_HIDE_FROM_ABI void destroy(pointer __p) { __p->~_Tp(); }

For transparency, I'm an ecologist but I've never worked with biomechanical data and I don't know anything about EZC3D as a C++ software library, but I do know C++ and Rcpp. Looking at your Makevars, I see that you don't have any flags indicating which version of C++ you want to be using. Adding flags might help prevent these warnings. I don't know which version of C++ is compatible with the ezc3d library but apparently it's not C++17 so maybe flag a different C++ version there.

When I ran a package check locally, I did get a note about the object files in the ezc3d folder in /src, but it's just a note and I assume this is the cost of doing business when you include a C++ library inside your R package.

Bravo on the test suite; it's very thorough and all tests passed on my machine.

I think that for the `c3d_data` and `c3d_convert` documentation you should be clear about the column names that the user can expect in the output data frames of each format, as well as column types.

For the example data, I think you should give more details about the data and how they were collected. Which lab collected the data? When? Is there a paper about it people can read?

For `c3d_setdata` I think you should make it clear whether you can write any data from the `c3d_data` function (wide, long, or longest). If I were using this function I would wonder if all three formats work equally well here.

The package documentation at ?c3dr should include a link to the EZC3D library website.

The vignette doesn't have a title other than "Vignette's Title" so you should probably change that. I appreciate that the vignette explained things on a very basic level since I, as stated, don't know anything about biomechanical data. You really broke it down for the novice user.

#### Conformation with rOpenSci guidelines

Looking at the code itself, `c3d_setdata` doesn't have input checking for the newdata or newanalog arguments. The `read` and `write` functions don't have input checking either (tip: easier and better to do in the R wrapper than the C++).
