### 2025-03-27 #686 @astamm
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☐ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software have been confirmed.
-   ☒ **Performance:** Any performance claims of the software have been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☐ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 4

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

#### Overall comments

I like the idea of a light package doing one specific task, which makes it easy to maintain and use. The package is well documented and the vignette is a good entry point. I appreciate all the efforts that the author made to keep the dependency trail as short as possible. The package is well structured and the code is clean and easy to read. I have some suggestions for improvement.

Similarly to @japilo’s review, I did not check the *installation* item for the same reason. I experimented by adding `CXX_STD=CXX11` to the `Makevars` but that does not help in removing those two warnings. I also think it is a minor thing as CRAN is not complaining about it. I would suggest to keep it as is for now but we can dig into it if you want.

Also similarly to @japilo’s review, I did not check the *Packaging guidelines* item for the same reason. The code could be improved by more systematically check for input validity and throw errors when necessary.

#### C++ code

-   I would avoid using `using namespace Rcpp` and other `using namespace ezc3d` as it makes it hard to track where functions are coming from. Instead, you could use `Rcpp::` or `ezc3d::` where necessary. It is just a suggestion of course.

-   Why have header files under `inst/include` instead of directly in the `src/` directory? Do you plan to have potentially downstream dependencies that would need to access the C++ API directly? I think that would be nice but we should demonstrate that it works. Have you tried it by creating a fake package which would use `c2dr` as a `LinkingTo` dependency? If this is outside the scope of this package, I would suggest moving the headers to the `src/` directory.

-   To avoid the need to maintain a manually curated list of source files, you might consider section [**1.2.1.3 Compiling in subdirectories**](https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Compiling-in-subdirectories) in the WRE manual for a more organized way to include headers and source files. In particular, if the EZC3D has a proper `Makefile` you could include all sources in `src/ezc3d/` and have a `Makevars` file that look like:

    PKG_LIBS = -ezc3d/lib -lezc3d $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)

    $(SHLIB): ezc3d/lib/libezc3d.a

    ezc3d/lib/libsdp.a:
            @(cd ezc3d/lib && $(MAKE) libezc3d.a \
              CC="$(CC)" CFLAGS="$(CFLAGS) $(CPICFLAGS)" AR="$(AR)" RANLIB="$(RANLIB)")

-   There are specific recommendations about how we should build against external software sources. Quoting [*Writing R Extensions (WRE)*](https://cran.r-project.org/doc/FAQ/R-exts.html#Writing-R-Extensions):

> N.B.: Include paths are preprocessor options, not compiler options, and must be set in PKG\_CPPFLAGS as otherwise platform-specific paths (e.g. ‘-I/usr/local/include’) will take precedence. PKG\_CPPFLAGS should contain ‘-I’, ‘-D’, ‘-U’ and (where supported) ‘-include’ and ‘-pthread’ options: everything else should be a compiler flag. The order of flags matters, and using ‘-I’ in PKG\_CFLAGS or PKG\_CXXFLAGS has led to hard-to-debug platform-specific errors.

So I would recommend using `PKG_CPPFLAGS` instead of `PKG_CFLAGS` in `src/Makevars`.

-   There is the following `NOTE` in the CRAN check (also spotted by @maelle and @japilo):

    #> ❯ checking if this is a source package ... NOTE
    #>   Found the following apparent object files/libraries:
    #>     src/ezc3d/Analogs.o src/ezc3d/AnalogsInfo.o
    #>     src/ezc3d/AnalogsSubframe.o src/ezc3d/Channel.o src/ezc3d/Data.o
    #>     src/ezc3d/DataStartInfo.o src/ezc3d/Frame.o src/ezc3d/Group.o
    #>     src/ezc3d/Header.o src/ezc3d/Parameter.o src/ezc3d/Parameters.o
    #>     src/ezc3d/Point.o src/ezc3d/Points.o src/ezc3d/PointsInfo.o
    #>     src/ezc3d/Rotation.o src/ezc3d/Rotations.o src/ezc3d/RotationsInfo.o
    #>     src/ezc3d/RotationsSubframe.o src/ezc3d/ezc3d.o
    #>     src/ezc3d/math/Matrix.o src/ezc3d/math/Matrix33.o
    #>     src/ezc3d/math/Matrix44.o src/ezc3d/math/Matrix66.o
    #>     src/ezc3d/math/Vector3d.o src/ezc3d/math/Vector6d.o
    #>     src/ezc3d/modules/ForcePlatforms.o
    #>   Object files/libraries should not be included in a source package.

This is because we need to clean things after the compilation. Actually, the R compilation mechanisms do that for us under `src/` but not automatically in its subdirectories. You can improve your `Makevars` with something along those lines:

    PKG_CPPFLAGS = -I../inst/include/ -I../inst/include/ezc3d/

    ##include files manually
    C3DRSOURCES = \
        ./ezc3d/Analogs.cpp \
        ./ezc3d/AnalogsInfo.cpp \
        ./ezc3d/AnalogsSubframe.cpp \
        ./ezc3d/Channel.cpp \
        ./ezc3d/Data.cpp \
        ./ezc3d/DataStartInfo.cpp \
        ./ezc3d/ezc3d.cpp \
        ./ezc3d/Frame.cpp \
        ./ezc3d/Group.cpp \
        ./ezc3d/Header.cpp \
        ./ezc3d/Parameter.cpp \
        ./ezc3d/Parameters.cpp \
        ./ezc3d/Point.cpp \
        ./ezc3d/Points.cpp \
        ./ezc3d/PointsInfo.cpp \
        ./ezc3d/Rotation.cpp \
        ./ezc3d/Rotations.cpp \
        ./ezc3d/RotationsInfo.cpp \
        ./ezc3d/RotationsSubframe.cpp \
        ./ezc3d/math/Matrix.cpp \
        ./ezc3d/math/Matrix33.cpp \
        ./ezc3d/math/Matrix44.cpp \
        ./ezc3d/math/Matrix66.cpp \
        ./ezc3d/math/Vector3d.cpp \
        ./ezc3d/math/Vector6d.cpp \
        ./ezc3d/modules/ForcePlatforms.cpp

    WRAPSOURCES = \
            read.cpp \
            write.cpp

    PKGSOURCES = RcppExports.cpp

    SOURCES = $(WRAPSOURCES) $(PKGSOURCES) $(C3DRSOURCES)
    OBJECTS = $(SOURCES:.cpp=.o)

    all: clean

    clean:
        rm -f $(SHLIB) $(OBJECTS)

-   It is good practice when bundling external sources to check the package for memory leaks and sorts. This can be achieved using R-hub, in particular against the `clang-ubsan, R-devel (2025-03-26 r88060), Ubuntu  22.04.5 LTS` configuration. Have you tried this?

#### R code

-   I noticed that `c3d` objects have `NaN B` size in RStudio environment pane after running for example `d <- c3d_read(c3d_example())`.

-   You might want to consider also a `print.c3d_analog()` method to avoid overflow of data frame printing in the console. Or make it a `tibble` instead of a `data.frame` but that will make the dependency trail longer.

-   The same goes for the `c3d_data` class.

#### Test suite

The test suite is well written and covers a good range of cases, congrats!

Following your own philosphy (also in your responses to @maelle) to keep the dependency trail as short as possible, I would suggest using [**tinytest**](https://github.com/markvanderloo/tinytest) instead of **testthat**. **tinytest** is a very light package that does not depend on any other package. Its only downside is that lack of support for snapshots, but there is now a package called [**tinysnapshot**](https://github.com/vincentarelbundock/tinysnapshot) that can be used to fill this gap.

#### Small comments

-   In the `Description` field of the `DESCRIPTION` file, you write *A wrapper* instead of *An wrapper*.

-   In the documentation of `c3d_example()`, I think you forgot the verb in one sentence:

> The test data file **is** a short recording of human walking using a full-body model.
