### 2025-03-28 #686 @maelle
Merci @astamm for your thorough review!

@smnnlt Now both reviews are in 🎉 https://devguide.ropensci.org/softwarereview_author.html#the-review-process
