### 2025-03-28 #686 @smnnlt
@japilo @astamm Thank you very much for your reviews! I think these are great suggestions to improve the package. I will try to make all the changes and respond to your comments within the next week.
