### 2025-04-03 #686 @smnnlt
Hi @japilo, @astamm, @maelle. Thank you very much again for your comments. Since this is my first time to really use C++ and its integration into R, I am happy to learn many new things in this regard. Your comments on the R code and the documentation were also very helpful. I will first make some general remarks before responding to your review point-by-point in the next messages.

Both of you noted compilation warnings during the installation process. I could not replicate the warnings you had (I may have a different compiler, but it should also use C++17). I did indeed find some other compilation warnings I could eliminate by adjusting the source code (https://github.com/smnnlt/c3dr/commit/ac578c7bb0169d9e172d36165b5032f542d344a7, https://github.com/smnnlt/c3dr/commit/ef266d50fac3413e102ddf6e40288250cea84835).

I would really like to systematically test different compilers on rhub. But I discovered that rhub currently does not support Quarto vignettes (https://github.com/r-hub/actions/issues/33). When I manually modify the action workflows I run into related problems on all non-Windows devices (https://github.com/r-hub/actions/issues/25). This means at the moment I cannot run the Linux-based rhub workflows for different C++ compilers. I definitely want to get that work before creating the next c3dr release, either by waiting for the rhub update or by further modifying my fork of the workflows.

R-CMD check of my package after the changes in response to the review comments currently shows no errors and warnings, but one note on Ubuntu:
```
 * checking installed package size ... INFO
   installed size is 11.2Mb
   sub-directories of 1Mb or more:
     extdata   1.2Mb
     libs      9.5Mb
```
But I think this should not be a problem.

See below for my detailed responses.
