### 2025-04-03 #686 @smnnlt
### Response to @japilo

> ### Review Comments
> #### Tests and Checks
>
> I did not mark that installation succeeds as documented because, while the package did install successfully, there were several warnings generated during the Rcpp installation. The warnings looked like this:
>
> > 168 |   _LIBCPP_DEPRECATED_IN_CXX17 _LIBCPP_HIDE_FROM_ABI void destroy(pointer __p) { __p->~_Tp(); }
>
> For transparency, I'm an ecologist but I've never worked with biomechanical data and I don't know anything about EZC3D as a C++ software library, but I do know C++ and Rcpp. Looking at your Makevars, I see that you don't have any flags indicating which version of C++ you want to be using. Adding flags might help prevent these warnings. I don't know which version of C++ is compatible with the ezc3d library but apparently it's not C++17 so maybe flag a different C++ version there.

As stated above, I could not replicate the exact warning you had, but found other warnings I could remove. I want to systematically test for warnings when using different compilers, but I first need to get rhub working for the package.

> When I ran a package check locally, I did get a note about the object files in the ezc3d folder in /src, but it's just a note and I assume this is the cost of doing business when you include a C++ library inside your R package.

I could remove the note by including the object files in .Rbuildignore (https://github.com/smnnlt/c3dr/commit/4ac4a7bc94016d04f3c5a391934c5635a6dfc7c1)

> Bravo on the test suite; it's very thorough and all tests passed on my machine.

Thank you very much!
> #### Documentation
>
> I think that for the `c3d_data` and `c3d_convert` documentation you should be clear about the column names that the user can expect in the output data frames of each format, as well as column types.

I agree, more information on the different data formats is needed. I modified and restructured the documentation (https://github.com/smnnlt/c3dr/commit/fb474b1bec7537fff170c0bb7345f60722751617).

> For the example data, I think you should give more details about the data and how they were collected. Which lab collected the data? When? Is there a paper about it people can read?

To be honest, I do not have much more information on the example data. I use a file from the ezc3d test suite which does not provide any background information about the measurements. I chose the file because it is fairly small, but captures all important aspects of c3d files (point data, analog data, and force platform data all available). I added more information on the measurement system of the recording in the documentation (https://github.com/smnnlt/c3dr/commit/ce57822ed806e84bc39b17fbe2e022fc8880247a).

> For `c3d_setdata` I think you should make it clear whether you can write any data from the `c3d_data` function (wide, long, or longest). If I were using this function I would wonder if all three formats work equally well here.

Added (https://github.com/smnnlt/c3dr/commit/554cdc00d146bdece8c1f91c0ab487373e81205f)
>
> The package documentation at ?c3dr should include a link to the EZC3D library website.

The main package documentation was mainly generated from the Description file, which I does not include a link. So I added a new section to the page (https://github.com/smnnlt/c3dr/commit/995fbcc104abaaca97ba7c28888038d3eec6109b).
>
> The vignette doesn't have a title other than "Vignette's Title" so you should probably change that. I appreciate that the vignette explained things on a very basic level since I, as stated, don't know anything about biomechanical data. You really broke it down for the novice user.

I am very happy that the vignette fulfilled its purpose for you. I added the title for indexing (https://github.com/smnnlt/c3dr/commit/3a649db22774f9c1c3b30d0e4265219043717b5e).
> #### Conformation with rOpenSci guidelines
>
> Looking at the code itself, `c3d_setdata` doesn't have input checking for the newdata or newanalog arguments. The `read` and `write` functions don't have input checking either (tip: easier and better to do in the R wrapper than the C++).

The input validation for `c3d_setdata()` is one level down in the internal `c3dr:::create_newdata()` and `c3dr:::create_newanalog()`. To make this easier to spot for developer, I added a comment (https://github.com/smnnlt/c3dr/commit/da57465429433398dd83bdefd0f59e764dcfb990). Let me know if this input validation should be directly in the top-level function.

I added basic input validation for `c3d_read()` and `c3d_write()` in the R wrappers (https://github.com/smnnlt/c3dr/commit/03adbf87332e40cb8611d27ad96cd6f9f5e5b03d, https://github.com/smnnlt/c3dr/commit/e42fa77655550eae8140e61ef57a21f3d5f828fb).

Thank you very much for your review!
