### 2025-04-03 #686 @smnnlt
### Response to @astamm

> ### Review Comments
> #### Overall comments
>
> I like the idea of a light package doing one specific task, which makes it easy to maintain and use. The package is well documented and the vignette is a good entry point. I appreciate all the efforts that the author made to keep the dependency trail as short as possible. The package is well structured and the code is clean and easy to read. I have some suggestions for improvement.

Thank you very much for your kind words and the valuable suggestions!
>
> Similarly to [@japilo](https://github.com/japilo)’s review, I did not check the _installation_ item for the same reason. I experimented by adding `CXX_STD=CXX11` to the `Makevars` but that does not help in removing those two warnings. I also think it is a minor thing as CRAN is not complaining about it. I would suggest to keep it as is for now but we can dig into it if you want.

As stated in my other messages, I could not replicate the exact error and still have to systematically test the package for different compilers.
>
> Also similarly to [@japilo](https://github.com/japilo)’s review, I did not check the _Packaging guidelines_ item for the same reason. The code could be improved by more systematically check for input validity and throw errors when necessary.

See my response to @japilo (https://github.com/smnnlt/c3dr/commit/da57465429433398dd83bdefd0f59e764dcfb990, https://github.com/smnnlt/c3dr/commit/03adbf87332e40cb8611d27ad96cd6f9f5e5b03d)
> #### C++ code
>
>  I would avoid using `using namespace Rcpp` and other `using namespace ezc3d` as it makes it hard to track where functions are coming from. Instead, you could use `Rcpp::` or `ezc3d::` where necessary. It is just a suggestion of course.

Thank you for the suggestion, I updated the .cpp files (https://github.com/smnnlt/c3dr/commit/f2affad163de022b72e5432a21fdb23e1f78c239)
>
>  Why have header files under `inst/include` instead of directly in the `src/` directory? Do you plan to have potentially downstream dependencies that would need to access the C++ API directly? I think that would be nice but we should demonstrate that it works. Have you tried it by creating a fake package which would use `c3dr` as a `LinkingTo` dependency? If this is outside the scope of this package, I would suggest moving the headers to the `src/` directory.

Actually I do not plan to build further downstream dependencies and do not think anyone else will. So I moved all header files to `src` (https://github.com/smnnlt/c3dr/commit/460aa33c1d7cf56378b29b6c953217c859686ec4)
>
>   To avoid the need to maintain a manually curated list of source files, you might consider section [**1.2.1.3 Compiling in subdirectories**](https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Compiling-in-subdirectories) in the WRE manual for a more organized way to include headers and source files. In particular, if the EZC3D has a proper `Makefile` you could include all sources in `src/ezc3d/` and have a `Makevars` file that look like:
>
> ```
> PKG_LIBS = -ezc3d/lib -lezc3d $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
>
> $(SHLIB): ezc3d/lib/libezc3d.a
>
> ezc3d/lib/libsdp.a:
>         @(cd ezc3d/lib && $(MAKE) libezc3d.a \
>           CC="$(CC)" CFLAGS="$(CFLAGS) $(CPICFLAGS)" AR="$(AR)" RANLIB="$(RANLIB)")
> ```

For now, I do not expect any changes in the list of source files in the future and I would need to create a new `Makefile` for EZC3D to make this method work. So I do not really see a major benefit of making this change, but admittedly I could definitely make the modification with some more work. So let me know if you believe these changes are essential.
>
>   There are specific recommendations about how we should build against external software sources. Quoting [_Writing R Extensions (WRE)_](https://cran.r-project.org/doc/FAQ/R-exts.html#Writing-R-Extensions):
>
>
> > N.B.: Include paths are preprocessor options, not compiler options, and must be set in PKG_CPPFLAGS as otherwise platform-specific paths (e.g. ‘-I/usr/local/include’) will take precedence. PKG_CPPFLAGS should contain ‘-I’, ‘-D’, ‘-U’ and (where supported) ‘-include’ and ‘-pthread’ options: everything else should be a compiler flag. The order of flags matters, and using ‘-I’ in PKG_CFLAGS or PKG_CXXFLAGS has led to hard-to-debug platform-specific errors.
>
> So I would recommend using `PKG_CPPFLAGS` instead of `PKG_CXXFLAGS` in `src/Makevars`.

Done (https://github.com/smnnlt/c3dr/commit/e2fd0e3a4ed6e19652027b7cbfec5374b08e6188).
>
>   There is the following `NOTE` in the CRAN check (also spotted by [@maelle](https://github.com/maelle) and [@japilo](https://github.com/japilo)):
>
> ```
> #> ❯ checking if this is a source package ... NOTE
> #>   Found the following apparent object files/libraries:
> #>     src/ezc3d/Analogs.o src/ezc3d/AnalogsInfo.o
> #>     src/ezc3d/AnalogsSubframe.o src/ezc3d/Channel.o src/ezc3d/Data.o
> #>     src/ezc3d/DataStartInfo.o src/ezc3d/Frame.o src/ezc3d/Group.o
> #>     src/ezc3d/Header.o src/ezc3d/Parameter.o src/ezc3d/Parameters.o
> #>     src/ezc3d/Point.o src/ezc3d/Points.o src/ezc3d/PointsInfo.o
> #>     src/ezc3d/Rotation.o src/ezc3d/Rotations.o src/ezc3d/RotationsInfo.o
> #>     src/ezc3d/RotationsSubframe.o src/ezc3d/ezc3d.o
> #>     src/ezc3d/math/Matrix.o src/ezc3d/math/Matrix33.o
> #>     src/ezc3d/math/Matrix44.o src/ezc3d/math/Matrix66.o
> #>     src/ezc3d/math/Vector3d.o src/ezc3d/math/Vector6d.o
> #>     src/ezc3d/modules/ForcePlatforms.o
> #>   Object files/libraries should not be included in a source package.
> ```
>
> This is because we need to clean things after the compilation. Actually, the R compilation mechanisms do that for us under `src/` but not automatically in its subdirectories. You can improve your `Makevars` with something along those lines:
>
> ```
> PKG_CPPFLAGS = -I../inst/include/ -I../inst/include/ezc3d/
>
> ##include files manually
> C3DRSOURCES = \
>     ./ezc3d/Analogs.cpp \
>     ./ezc3d/AnalogsInfo.cpp \
>     ./ezc3d/AnalogsSubframe.cpp \
>     ./ezc3d/Channel.cpp \
>     ./ezc3d/Data.cpp \
>     ./ezc3d/DataStartInfo.cpp \
>     ./ezc3d/ezc3d.cpp \
>     ./ezc3d/Frame.cpp \
>     ./ezc3d/Group.cpp \
>     ./ezc3d/Header.cpp \
>     ./ezc3d/Parameter.cpp \
>     ./ezc3d/Parameters.cpp \
>     ./ezc3d/Point.cpp \
>     ./ezc3d/Points.cpp \
>     ./ezc3d/PointsInfo.cpp \
>     ./ezc3d/Rotation.cpp \
>     ./ezc3d/Rotations.cpp \
>     ./ezc3d/RotationsInfo.cpp \
>     ./ezc3d/RotationsSubframe.cpp \
>     ./ezc3d/math/Matrix.cpp \
>     ./ezc3d/math/Matrix33.cpp \
>     ./ezc3d/math/Matrix44.cpp \
>     ./ezc3d/math/Matrix66.cpp \
>     ./ezc3d/math/Vector3d.cpp \
>     ./ezc3d/math/Vector6d.cpp \
>     ./ezc3d/modules/ForcePlatforms.cpp
>
> WRAPSOURCES = \
>         read.cpp \
>         write.cpp
>
> PKGSOURCES = RcppExports.cpp
>
> SOURCES = $(WRAPSOURCES) $(PKGSOURCES) $(C3DRSOURCES)
> OBJECTS = $(SOURCES:.cpp=.o)
>
> all: clean
>
> clean:
>     rm -f $(SHLIB) $(OBJECTS)
> ```

I have removed the note by modifying `.Rbuildignore` (https://github.com/smnnlt/c3dr/commit/4ac4a7bc94016d04f3c5a391934c5635a6dfc7c1). I restructured the `Makevars` based on your suggestions (https://github.com/smnnlt/c3dr/commit/5374c2286a80d26f7ee90787fa4521b7e9e99b56). I did not include the object cleaning, as WRE states (https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Using-Makevars):

>An unfortunately common error is to have
> `all: $(SHLIB) clean`
> which asks make to clean in parallel with compiling the code. Not only does this lead to hard-to-debug installation errors, it wipes out all the evidence of any error (from a parallel make or not). It is much better to leave cleaning to the end user using the facilities in the previous paragraph.

But maybe I misunderstood something here?

>
>   It is good practice when bundling external sources to check the package for memory leaks and sorts. This can be achieved using R-hub, in particular against the `clang-ubsan, R-devel (2025-03-26 r88060), Ubuntu  22.04.5 LTS` configuration. Have you tried this?

This is a great suggestion. Unfortunately, as described in my general comments, I currently have problems to get rhub running because of the quarto vignette.
>
> #### R code
>
>   I noticed that `c3d` objects have `NaN B` size in RStudio environment pane after running for example `d <- c3d_read(c3d_example())`.

I tried this out, and it seems that all object in the R environment that are larger than 1 MB are displayed with size `NaN`. Maybe this is a bug or deliberate feature of RStudio, but it seems to be unrelated to the package.
>
>   You might want to consider also a `print.c3d_analog()` method to avoid overflow of data frame printing in the console. Or make it a `tibble` instead of a `data.frame` but that will make the dependency trail longer.
>
>  The same goes for the `c3d_data` class.

Yes, I also considered this during the development. For me, I prefer objects with a simple structure (e.g., data frames) to print in a way that shows the user the data they are actually dealing with (the opposite would be complex data structures with dedicated accessor functions, where I am totally fine with the printing output looking very different to the object and its structure, e.g., `print.c3d()`). So printing for point or analog data would need to be something very similar to `tibble` that still shows the data and not just a data summary. I am not a fan of using `tibble` in the package because of (1) the extra dependency, as you said, (2) other modifications `tibble` makes to a `data.frame`, for example changing single column indexing with `[]`. I have to admit those are personal reservations against `tibble`, and I do not really see enough benefits of manually creating a print function resembling `tibble` just to avoid overflow. But if you feel this is important, or have any suggestions for easy dependency-free print methods for this package, I am happy to implement them.
>
> #### Test suite
>
> The test suite is well written and covers a good range of cases, congrats!

Thank you very much!
>
> Following your own philosphy (also in your responses to [@maelle](https://github.com/maelle)) to keep the dependency trail as short as possible, I would suggest using [**tinytest**](https://github.com/markvanderloo/tinytest) instead of **testthat**. **tinytest** is a very light package that does not depend on any other package. Its only downside is that lack of support for snapshots, but there is now a package called [**tinysnapshot**](https://github.com/vincentarelbundock/tinysnapshot) that can be used to fill this gap.

Thank you for pointing out to `tinytest` and `tinysnapshot`. I have not used them before, but I really like their approach of making the testing infrastructure simpler. That said, there are some minor drawbacks I discovered when I translated all tests to `tinytest` (which generally worked extremely well). Some of the improvements in response to suggestions made by @maelle in her review earlier would need to be reverted, e.g., because `tinytest` cannot explicitly test for classes, and because code cannot be capsuled within each test. But the biggest drawback, that lead me to deciding against `tinytest` was that it cannot be integrated into the `devtools` workflow (and this does not seem to change anytime in the future, see https://github.com/r-lib/devtools/issues/2361). But maybe I overlooked something here? Since I heavily use a `devtools`-based workflow for development, I do not see a benefit of replacing `testthat` at the moment, given that the extra dependency only concerns `Suggests` and thus is not relevant to most package users.
> #### Small comments
>
>     * In the `Description` field of the `DESCRIPTION` file, you write _An wrapper_ instead of _A wrapper_.
>
>     * In the documentation of `c3d_example()`, I think you forgot the verb in one sentence:
>
> > The test data file **contains** a short recording of human walking using a full-body model.

Fixed! (https://github.com/smnnlt/c3dr/commit/4f9c10ff63d6cac0f6ee6a376fa486def79b09ec)

Thank you very much for your review!
