### 2025-04-07 #686 @astamm
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 5

Thank you very much @smnnlt for your detailed and precise answer! I have not yet checked the final approval box as I would like some more comments on some items I brought up. See below.

> Thank you for the suggestion, I updated the .cpp files ([smnnlt/c3dr@f2affad](https://github.com/smnnlt/c3dr/commit/f2affad163de022b72e5432a21fdb23e1f78c239))

I should not have been that drastic. It is a good practice to do as I suggested. However, some files are generated automatically by **Rcpp** such as `RcppExports.cpp` in which `using namespace Rcpp;` is used. This file should not be edited by hand as in any case it will be regenerated without `Rcpp::` when you rebuild the package from scratch.

> Actually I do not plan to build further downstream dependencies and do not think anyone else will. So I moved all header files to `src` ([smnnlt/c3dr@460aa33](https://github.com/smnnlt/c3dr/commit/460aa33c1d7cf56378b29b6c953217c859686ec4))

If this is the intended use, that's fine. But I am a bit surprised that you do not predict your package will be used down the line by other people working on motion data. It could be a nice feature that the C++ interface of your package (i.e. reader and writer) could be used by R package developers who would use C++ behind the scenes. But not my call :-).

> For now, I do not expect any changes in the list of source files in the future and I would need to create a new `Makefile` for EZC3D to make this method work. So I do not really see a major benefit of making this change, but admittedly I could definitely make the modification with some more work. So let me know if you believe these changes are essential.

I agree with you.

> Done ([smnnlt/c3dr@e2fd0e3](https://github.com/smnnlt/c3dr/commit/e2fd0e3a4ed6e19652027b7cbfec5374b08e6188)).

Note that if you move all headers to `src/`, you do not need `PKG_CPPFLAGS = -I.` in `Makevars` I think.

> I have removed the note by modifying `.Rbuildignore` ([smnnlt/c3dr@4ac4a7b](https://github.com/smnnlt/c3dr/commit/4ac4a7bc94016d04f3c5a391934c5635a6dfc7c1)). I restructured the `Makevars` based on your suggestions ([smnnlt/c3dr@5374c22](https://github.com/smnnlt/c3dr/commit/5374c2286a80d26f7ee90787fa4521b7e9e99b56)). I did not include the object cleaning, as WRE states (https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Using-Makevars)

Nice catch to put the `*.o` files in `.Rbuildignore` but it is still good practice to clean up after ourselves :-). The thing WRE warns us against is that, when you do `all: $(SHLIB) clean`, you basically ask simultaneously to build the sources and clean the object files at the same time. This is risky especially on some platforms where R executes these targets in parallel. What I proposed was simply `all: clean` but it is risky too because we do not specify where and when should the sources be compiled (which is what the $(SHLIB) target is about).

Anyway, WRE says to refer to the facilities described in the previous paragraph to handle cleaning. In particular, it is quite simple as it only requires to add a `cleanup` bash file at the root of your package with something along those lines:

```bash
#!/bin/sh

rm -rf src/*.o src/ezc3d/*.o src/ezc3d/math/*.o src/ezc3d/modules/*.o src/*.so
```

You might have to run `chmod 755 cleanup` in a terminal to give proper read/write permissions to R.

> This is a great suggestion. Unfortunately, as described in my general comments, I currently have problems to get rhub running because of the quarto vignette.

I see progress on your repo so I assume you managed to get it to work, congrats!

> I tried this out, and it seems that all object in the R environment that are larger than 1 MB are displayed with size `NaN`. Maybe this is a bug or deliberate feature of RStudio, but it seems to be unrelated to the package.

Agreed.

> Yes, I also considered this during the development. For me, I prefer objects with a simple structure (e.g., data frames) to print in a way that shows the user the data they are actually dealing with (the opposite would be complex data structures with dedicated accessor functions, where I am totally fine with the printing output looking very different to the object and its structure, e.g., `print.c3d()`). So printing for point or analog data would need to be something very similar to `tibble` that still shows the data and not just a data summary. I am not a fan of using `tibble` in the package because of (1) the extra dependency, as you said, (2) other modifications `tibble` makes to a `data.frame`, for example changing single column indexing with `[]`. I have to admit those are personal reservations against `tibble`, and I do not really see enough benefits of manually creating a print function resembling `tibble` just to avoid overflow. But if you feel this is important, or have any suggestions for easy dependency-free print methods for this package, I am happy to implement them.

I don't have a strong opinion about this. I prefer short output that do not take pages in the console, which is what I like about tibbles. Maybe calling the `head()` function within the `print()` method could be a simple way to help with that.

Also, I forgot to mention that before, but there are two methods for controlling the output to the console: `format()` and `print()`. Usually, what you implemented goes in the body of the `format()` method and the `print()` method simply reads:

```r
print.c3d <- function(x, ...) {
  cat(format(x, ...), sep = "\n")
  invisible(x)
}
```

> Thank you for pointing out to `tinytest` and `tinysnapshot`. I have not used them before, but I really like their approach of making the testing infrastructure simpler. That said, there are some minor drawbacks I discovered when I translated all tests to `tinytest` (which generally worked extremely well). Some of the improvements in response to suggestions made by [@maelle](https://github.com/maelle) in her review earlier would need to be reverted, e.g., because `tinytest` cannot explicitly test for classes, and because code cannot be capsuled within each test. But the biggest drawback, that lead me to deciding against `tinytest` was that it cannot be integrated into the `devtools` workflow (and this does not seem to change anytime in the future, see [r-lib/devtools#2361](https://github.com/r-lib/devtools/issues/2361)). But maybe I overlooked something here? Since I heavily use a `devtools`-based workflow for development, I do not see a benefit of replacing `testthat` at the moment, given that the extra dependency only concerns `Suggests` and thus is not relevant to most package users.

I understand the point of view. Especially that now you have LLM-powered tools such as [**chores**](https://simonpcouch.github.io/chores/) and [**ensure**](https://simonpcouch.github.io/ensure/) that provides AI tools for building up the test suite automatically with **testthat** using LLM trained on purpose to be *testthat*-aware.

Side note about tinytest: there is an RStudio Addin which still facilitates the workflow.
