### 2025-04-09 #686 @japilo
## Reviewer Response

> I agree, more information on the different data formats is needed. I modified and restructured the documentation.

I think you did a great job here. I read the documentation and I now understand the format of the different data configurations. I also think your addition to `c3d_setdata` clarifies which data formats are supported.

> To be honest, I do not have much more information on the example data. I use a file from the ezc3d test suite which does not provide any background information about the measurements. I chose the file because it is fairly small, but captures all important aspects of c3d files (point data, analog data, and force platform data all available). I added more information on the measurement system of the recording in the documentation

Ah well, I suppose you can't do much about that. The documentation you provided is sufficient under the circumstances.

> The input validation for c3d_setdata() is one level down in the internal c3dr:::create_newdata() and c3dr:::create_newanalog()

Oh okay, thanks for that. Now that I can see the input validation, I do have some suggestions. You're still missing validation for the `nperframe` argument. It should presumably be a positive integer. You may also want to have a check that the data have more than 0 rows. For `c3d_read` and `c3d_write`, you can improve the validation of the file path by checking it with `file.exists()`. I would like to see these changes made before I approve the package.

#### Final approval (post-review)

Estimated hours spent reviewing: 4.5
