### 2025-04-14 #686 @smnnlt
Hi @astamm

Thank you very much for your response!

> I should not have been that drastic. It is a good practice to do as I suggested. However, some files are generated automatically by **Rcpp** such as `RcppExports.cpp` in which `using namespace Rcpp;` is used. This file should not be edited by hand as in any case it will be regenerated without `Rcpp::` when you rebuild the package from scratch.

I did not modify `RcppExports.cpp` by hand. All changes were made automatically because of the manual changes in `read.cpp` and `write.cpp`.

> If this is the intended use, that's fine. But I am a bit surprised that you do not predict your package will be used down the line by other people working on motion data. It could be a nice feature that the C++ interface of your package (i.e. reader and writer) could be used by R package developers who would use C++ behind the scenes. But not my call :-).
>
Maybe a was a bit too firm here. I do expect that others will build up on my code. I just do expect that most will do this by accessing the R interface instead of the C++ interface within R. I have created a demonstration package to show how a downstream dependency with `c3dr` in `Imports` works: https://github.com/smnnlt/c3drRevdep
>
> > Done ([smnnlt/c3dr@e2fd0e3](https://github.com/smnnlt/c3dr/commit/e2fd0e3a4ed6e19652027b7cbfec5374b08e6188)).
>
> Note that if you move all headers to `src/`, you do not need `PKG_CPPFLAGS = -I.` in `Makevars` I think.

Okay (https://github.com/smnnlt/c3dr/pull/3/commits/d214d9a6911ec2c7ceba09e6ce1f26814c8e4abc)

> Nice catch to put the `*.o` files in `.Rbuildignore` but it is still good practice to clean up after ourselves :-). The thing WRE warns us against is that, when you do `all: $(SHLIB) clean`, you basically ask simultaneously to build the sources and clean the object files at the same time. This is risky especially on some platforms where R executes these targets in parallel. What I proposed was simply `all: clean` but it is risky too because we do not specify where and when should the sources be compiled (which is what the $(SHLIB) target is about).
>
> Anyway, WRE says to refer to the facilities described in the previous paragraph to handle cleaning. In particular, it is quite simple as it only requires to add a `cleanup` bash file at the root of your package with something along those lines:
>
> #!/bin/sh
>
> rm -rf src/*.o src/ezc3d/*.o src/ezc3d/math/*.o src/ezc3d/modules/*.o src/*.so
>
> You might have to run `chmod 755 cleanup` in a terminal to give proper read/write permissions to R.

Good point! I added a `cleanup` and `cleanup.win` file (https://github.com/smnnlt/c3dr/pull/3/commits/02489f04ea43aa92070b737205122eb89c8ef8b6, https://github.com/smnnlt/c3dr/pull/3/commits/514eb35aa1aef4838ea6463d18f12b4736653967). I could locally verify on Windows that the `.o` files are removed after the build, and on other platforms R-CMD does not complain at least.

> > This is a great suggestion. Unfortunately, as described in my general comments, I currently have problems to get rhub running because of the quarto vignette.
>
> I see progress on your repo so I assume you managed to get it to work, congrats!

Not really, r-hub caused me a lot of trouble lately. My current workaround is to have a separate branch in which I replaced the Quarto vignette with a RMarkdown version. This way I could get r-hub running on all platforms. R-CMD worked well, but clang-ubsan failed, from what I understand because of problems in the underlying ezc3d source code (https://github.com/smnnlt/c3dr/actions/runs/14442539771). Because R-CMD is not complaining on all important platform, I am a bit unsure on how to continue.

> I don't have a strong opinion about this. I prefer short output that do not take pages in the console, which is what I like about tibbles. Maybe calling the `head()` function within the `print()` method could be a simple way to help with that.

I gave this a second thought and experimented with adding a `header()` call to print. The problem is that we are often dealing with data frames with many columns here, so restricting the row number alone does not prevent overflow. To have a clean print output I would also have to cut the columns, which would subsequently require additional notes to the user about the omitted data and how to access it. So in the end, this would require a pretty much `tibble`-like solution, just with a manual implementation. I do not see the benefit of doing this work at the moment, but again, this is my personal taste.

> Also, I forgot to mention that before, but there are two methods for controlling the output to the console: `format()` and `print()`. Usually, what you implemented goes in the body of the `format()` method and the `print()` method simply reads:
>
> print.c3d <- function(x, ...) {
>   cat(format(x, ...), sep = "\n")
>   invisible(x)
> }

Good idea, I added the method (https://github.com/smnnlt/c3dr/pull/3/commits/41d1922b07bbe6e77e997054de642a7862574f72)

> > Thank you for pointing out to `tinytest` and `tinysnapshot`. I have not used them before, but I really like their approach of making the testing infrastructure simpler. That said, there are some minor drawbacks I discovered when I translated all tests to `tinytest` (which generally worked extremely well). Some of the improvements in response to suggestions made by [@maelle](https://github.com/maelle) in her review earlier would need to be reverted, e.g., because `tinytest` cannot explicitly test for classes, and because code cannot be capsuled within each test. But the biggest drawback, that lead me to deciding against `tinytest` was that it cannot be integrated into the `devtools` workflow (and this does not seem to change anytime in the future, see [r-lib/devtools#2361](https://github.com/r-lib/devtools/issues/2361)). But maybe I overlooked something here? Since I heavily use a `devtools`-based workflow for development, I do not see a benefit of replacing `testthat` at the moment, given that the extra dependency only concerns `Suggests` and thus is not relevant to most package users.
>
> I understand the point of view. Especially that now you have LLM-powered tools such as [**chores**](https://simonpcouch.github.io/chores/) and [**ensure**](https://simonpcouch.github.io/ensure/) that provides AI tools for building up the test suite automatically with **testthat** using LLM trained on purpose to be _testthat_-aware.
>
> Side note about tinytest: there is an RStudio Addin which still facilitates the workflow.

Thank you for pointing out to the Addin. I think I'll stick to `testthat` for this project.
