### 2025-04-14 #686 @mpadge
@smnnlt The bot commands need to be separate there, so just repeat the bot command and nothing else, and include the link to the comment, and all should work
