### 2025-04-23 #686 @japilo
> Thanks [@smnnlt](https://github.com/smnnlt)!
>
> [@astamm](https://github.com/astamm) [@japilo](https://github.com/japilo) are you happy with [@smnnlt](https://github.com/smnnlt)'s responses?

Yes, I am happy with it. Informative error messages FTW!
