### 2025-04-23 #686 @astamm
> Thanks [@smnnlt](https://github.com/smnnlt)!
>
> [@astamm](https://github.com/astamm) [@japilo](https://github.com/japilo) are you happy with [@smnnlt](https://github.com/smnnlt)'s responses?

I am mostly happy with the replies.

I am though concerned about the rhub check failure against `clang-asan`. It spots two issues:

```bash
ezc3d/ezc3d.cpp:140:29: runtime error: 4.29497e+09 is outside the range of representable values of type 'int'
ezc3d/ezc3d.cpp:151:45: runtime error: 4.29497e+09 is outside the range of representable values of type 'unsigned int'
```

which I believe will upset CRAN maintainers who will ask you to fix.
