### 2025-04-23 #686 @smnnlt
@astamm Thank you for digging deeper into the ubsan failures.

> I am though concerned about the rhub check failure against `clang-asan`. It spots two issues:
>
> ezc3d/ezc3d.cpp:140:29: runtime error: 4.29497e+09 is outside the range of representable values of type 'int'
> ezc3d/ezc3d.cpp:151:45: runtime error: 4.29497e+09 is outside the range of representable values of type 'unsigned int'
>
> when running `d <- c3d_read(c3d_example())`, which I believe will upset CRAN maintainers who will ask you to fix.

I have modified `ezc3d.cpp` (https://github.com/smnnlt/c3dr/commit/8afa20d15b573d4fff1cd9ffdaa60a266d342763), so that now [clang-ubsan](https://github.com/smnnlt/c3dr/actions/runs/14622677951) and [clang-asan](https://github.com/smnnlt/c3dr/actions/runs/14622789543) run without error.
