### 2025-04-24 #686 @ropensci-review-bot
Approved! Thanks @smnnlt for submitting and @astamm, @japilo for your reviews! :grin:
